リファクタ: interface の移動
